Rename runtime.py to treewalker.py
